+++ androidApp/src/main/java/com/thinlineit/favorit_android/android/data/repository/PresentRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.thinlineit.favorit_android.android.data.Result
 import com.thinlineit.favorit_android.android.data.api.FundingApi
 import com.thinlineit.favorit_android.android.data.entity.PresentRequest
 import com.thinlineit.favorit_android.android.data.entity.PresentResult
-import com.thinlineit.favorit_android.android.ui.present.list.Present
+import com.thinlineit.favorit_android.android.data.entity.Present
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
